Remove graphql
