+++ packages/code-infra/package.json
@@ -61,7 +61,6 @@
     "@mui/internal-babel-plugin-resolve-imports": "workspace:*",
     "@next/eslint-plugin-next": "^15.5.4",
     "@octokit/auth-action": "^6.0.1",
-    "@octokit/graphql": "^9.0.1",
     "@octokit/rest": "^22.0.0",
     "@pnpm/find-workspace-dir": "^1000.1.3",
     "babel-plugin-optimize-clsx": "^2.6.2",
