+++ packages/code-infra/src/utils/github-types.mts
@@ -1,0 +1,9 @@
+export type AuthorAssocation =
+  | 'MEMBER'
+  | 'OWNER'
+  | 'COLLABORATOR'
+  | 'CONTRIBUTOR'
+  | 'FIRST_TIME_CONTRIBUTOR'
+  | 'FIRST_TIMER'
+  | 'NONE'
+  | 'MANNEQUIN';
